+++ src/components/Admin/Sidebar/Sidebar.tsx
@@ -37,9 +37,11 @@ const Sidebar = () => {
         setActive("Order");
         break;
       case "/admin/product":
+      case "/admin/product/diamond/detail/:id":
+      case "/admin/product/jewelry":
+      case "/admin/product/detail/:id":
       case "/admin/product/add/product":
       case "/admin/product/add/jewelry":
-      case "/admin/product/diamond":
       case "/admin/product/ring-setting":
       case "/admin/product/jewelry-type":
       case "/admin/product/material":
