+++ src/config/routes.ts
@@ -38,7 +38,9 @@ const routes = {
         cancelledOrder: '/admin/order/cancelled',
         orderDetail: '/admin/order/detail/:id',
         product: '/admin/product',
-        productDetail: '/admin/product/detail/:id',
+        diamondDetail: '/admin/product/diamond/detail/:id',
+        jewelryProduct: '/admin/product/jewelry',
+        jewelryDetail: '/admin/product/detail/:id',
         addProduct: '/admin/product/add/product',
         addJewelry: '/admin/product/add/jewelry',
         diamondProduct: '/admin/product/diamond',
